# Chacha20-Poly1305: use string instead of cstruct