--- a/src/native/poly1305-donna.c
+++ b/src/native/poly1305-donna.c
@@ -54,18 +54,18 @@ poly1305_update(poly1305_context *ctx, const unsigned char *m, size_t bytes) {
 }
 
 //stubs for OCaml
-CAMLprim value mc_poly1305_init (value ctx, value key, value off) {
-  poly1305_init ((poly1305_context *) Bytes_val(ctx), _ba_uint8_off(key, off));
+CAMLprim value mc_poly1305_init (value ctx, value key) {
+  poly1305_init ((poly1305_context *) Bytes_val(ctx), _st_uint8(key));
   return Val_unit;
 }
 
-CAMLprim value mc_poly1305_update (value ctx, value buf, value off, value len) {
-  poly1305_update ((poly1305_context *) Bytes_val(ctx), _ba_uint8_off(buf, off), Int_val(len));
+CAMLprim value mc_poly1305_update (value ctx, value buf, value len) {
+  poly1305_update ((poly1305_context *) Bytes_val(ctx), _st_uint8(buf), Int_val(len));
   return Val_unit;
 }
 
-CAMLprim value mc_poly1305_finalize (value ctx, value mac, value off) {
-  poly1305_finish ((poly1305_context *) Bytes_val(ctx), _ba_uint8_off(mac, off));
+CAMLprim value mc_poly1305_finalize (value ctx, value mac) {
+  poly1305_finish ((poly1305_context *) Bytes_val(ctx), Bytes_val(mac));
   return Val_unit;
 }