Performance improvement around 2.5x

Worth to gather some more statistics hereof, and the question how to move forward - I can see two ways:
- preserve old API, add some functions (adds noise)
- first reimplement stuff without Cstruct, preserve old API, then cut a release where the API doesn't include cstruct

WDYT?

main branch (28f8cde5ff3197e6383a935037725ab34ab32485)
       16:  11.686102 MB/s  (2318663 iters in 3.028 s)
       64:  44.914143 MB/s  (2248341 iters in 3.055 s)
      256:  111.408313 MB/s  (1464243 iters in 3.209 s)
     1024:  190.362636 MB/s  (530561 iters in 2.722 s)
     8192:  235.326363 MB/s  (81843 iters in 2.717 s)

this PR (with using the ad-hoc API enc_auth_str):
       16:  26.288283 MB/s  (5247580 iters in 3.046 s)
       64:  99.922308 MB/s  (4995212 iters in 3.051 s)
      256:  204.388046 MB/s  (2405283 iters in 2.873 s)
     1024:  278.060783 MB/s  (863503 iters in 3.033 s)
     8192:  294.746967 MB/s  (113552 iters in 3.010 s)