@hannesm 2024-02-25
note this could be split into pieces, such as introducing the xor_into_bytes function, then moving poly1305 to string, then adding `macl`, then moving chacha to string. on a separate note, in C we directly cast to `uint32_t *` to avoid shifting and reading byte-wise (this could also be done in a separate commit).