@palainp 2024-02-26
Neat! Thanks for that!
Here are my results (I confirm the x2 improvement on smallest sizes, and improvement everywhere!):
- head
       16:  10.119168 MB/s  (1347821 iters in 2.032 s)
       64:  38.610391 MB/s  (1286782 iters in 2.034 s)
      256:  94.388170 MB/s  (778949 iters in 2.015 s)
     1024:  144.286801 MB/s  (296597 iters in 2.007 s)
     8192:  179.547401 MB/s  (45952 iters in 1.999 s)
- this pr
       16:  20.138118 MB/s  (4008503 iters in 3.037 s)
       64:  74.277972 MB/s  (3603763 iters in 2.961 s)
      256:  142.762819 MB/s  (1758954 iters in 3.008 s)
     1024:  182.994248 MB/s  (562805 iters in 3.003 s)
     8192:  189.376331 MB/s  (72417 iters in 2.987 s)