@hannesm 2024-02-27
with @reynir suggested changes 1e1d179
* [chacha20-poly1305]
       16:  27.772872 MB/s  (5554249 iters in 3.052 s)
       64:  105.891640 MB/s  (5228603 iters in 3.014 s)
      256:  211.486396 MB/s  (2608886 iters in 3.012 s)
     1024:  281.579622 MB/s  (886396 iters in 3.074 s)
     8192:  299.032280 MB/s  (116109 iters in 3.033 s)